# openapi.yaml
openapi: 3.0.0
info:
  title: API E-commerce Sederhana
  description: API untuk mengelola pengguna, toko, produk, pesanan, dan transaksi.
  version: 1.0.0
servers:
  - url: /api/v1
    description: Server Lokal
components:
  schemas:
...
    OrderStatus:
      type: string
      enum: [pending, paid, shipped, delivered, cancelled, failed]
      description: Status pesanan.
...
    NewOrder:
      type: object
      required: [items, shipping_address]
      properties:
        shipping_address:
          type: string
          example: Jl. Sudirman No. 12, Jakarta Selatan
        items:
          type: array
          items:
            type: object
            required: [product_id, quantity]
            properties:
              product_id:
                type: integer
                example: 101
              quantity:
                type: integer
                example: 2
    Order: 
      type: object
      properties:
        id:
          type: string
          example: ORD-20251009-0001
          readOnly: true
        user_id:
          type: integer
          readOnly: true
        total_amount:
          type: number
          format: float
          readOnly: true
        status:
          $ref: '#/components/schemas/OrderStatus'
        created_at:
          type: string
          format: date-time
          readOnly: true
...
paths:
...
  /orders:
    get: # ✅ GET DAFTAR DITAMBAHKAN
      tags:
        - Orders
      summary: Daftar Pesanan Saya
      description: Mengambil daftar pesanan yang dibuat oleh pengguna yang terautentikasi.
      responses:
        '200':
          description: Daftar pesanan.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '401':
          description: Tidak terautentikasi.
    post:
      tags:
        - Orders
      summary: Membuat Pesanan Baru
      description: Pembeli membuat pesanan baru.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewOrder'
      responses:
        '201':
          description: Pesanan berhasil dibuat.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: ORD-20251009-0001
                  status:
                    $ref: '#/components/schemas/OrderStatus'
        '400':
          description: Stok tidak cukup atau input tidak valid.
  /orders/{orderId}:
    get: # ✅ GET DETAIL DITAMBAHKAN
      tags:
        - Orders
      summary: Mendapatkan Detail Pesanan
      description: Mengambil detail pesanan tertentu.
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detail pesanan.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order' 
        '404':
          description: Pesanan tidak ditemukan.
    delete: # ✅ DELETE DITAMBAHKAN (Pembatalan)
      tags:
        - Orders
      summary: Membatalkan Pesanan
      description: Membatalkan pesanan. (Hanya diizinkan jika status masih 'pending').
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Pesanan berhasil dibatalkan.
        '403':
          description: Pesanan tidak dapat dibatalkan.
        '404':
          description: Pesanan tidak ditemukan.
  /orders/{orderId}/status:
    put:
      tags:
        - Orders
      summary: Memperbarui Status Pesanan
      description: Memperbarui status pesanan.
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  $ref: '#/components/schemas/OrderStatus'
                  example: shipped
      responses:
        '200':
          description: Status pesanan berhasil diperbarui.
        '403':
          description: Tidak diizinkan.
          
  # -----------------------------------------------------
  # --- ENDPOINT ULASAN (reviews) ---
  # -----------------------------------------------------
...